\name{.get.function.dependencies}
\alias{.get.function.dependencies}
\title{Recursively determine the dependencies of a given function.}
\usage{.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")), src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
.get.function.dependencies(fun, packages, ignore.packages=c("base", getOption("defaultPackages")),
    src.root=getOption("src.root"), verbose=FALSE)
}
\description{Recursively determine the dependencies of a given function.}
\details{Not intended for public consumption. See \code{\link{get.function.dependencies}}\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


\code{.get.function.dependencies}: Not intended for public consumption. See \code{\link{get.function.dependencies}}


}
\value{a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


\code{.get.function.dependencies}: a list with these elements:
\item{my.functions}{a vector of all functions that are called}
\item{my.function.packages}{a vector indicating which package each function came from}
\item{other.packages}{a vector of other CRAN or BioConductor packages which are also needed}


}
\author{Mark Cowley}
\arguments{\item{fun}{either a function name, of the function's code}
\item{packages}{vector of packages to search for dependencies in (ie packages that you probably wrote)}
\item{ignore.packages}{a vector of packages to ignore. defaults to the list of pacakges loaded at startup.}
\item{verbose}{print the recursive dependency trace}
}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

\alias{.get.function.dependencies}

