\name{bioc.version}
\alias{bioc.version}
\title{What version of Bioconductor does this version of R use?}
\description{What version of Bioconductor does this version of R use?}
\details{Since at least R 2.5, Bioc has been 5 minor points behind, so
R 2.5 = bioc 2.0
R 2.12 = bioc 2.7 ...}
\value{The current version of bioc as a character string, eg "2.7"}
\author{Mark Cowley}
\examples{bioc.version()}

