% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_check.R
\name{check_internet}
\alias{check_internet}
\title{FUNCTION_TITLE}
\usage{
check_internet(url = "http://www.google.com")
}
\arguments{
\item{url}{PARAM_DESCRIPTION, Default: 'http://www.google.com'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
