% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_check.R
\name{check_libPaths}
\alias{check_libPaths}
\title{check .libPaths()}
\usage{
check_libPaths()
}
\value{
logical
}
\description{
checks if .libPaths only points to one folder inside running R
  installation. Will throw an error if this is not the case.
}
\examples{
\dontrun{
check_libPaths()
}
}
