% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_check.R
\name{check_no_of_tasks}
\alias{check_no_of_tasks}
\title{FUNCTION_TITLE}
\usage{
check_no_of_tasks(task_regex = "rsession\\\\.exe")
}
\arguments{
\item{task_regex}{PARAM_DESCRIPTION, Default: 'rsession\.exe'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
