% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_previous_library_loc.R
\name{find_previous_library_loc}
\alias{find_previous_library_loc}
\title{Find Package Library}
\usage{
find_previous_library_loc()
}
\value{
package library location
}
\description{
The function searches the system paths to find the location of the
previous R version's system library location.
It is recommended to \emph{not} use this function!
Rather, we recommend that a user simply open the previous version of R
and run \code{.libPaths()} to find the library location(s).
}
\examples{
find_previous_library_loc()
}
