% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_update.R
\name{get_Rversion_from_path}
\alias{get_Rversion_from_path}
\title{get R version from path}
\usage{
get_Rversion_from_path(path = Sys.getenv("R_HOME"))
}
\arguments{
\item{path, }{Default Sys.getenv('R_HOME')}
}
\value{
character vector 'X.X.X'
}
\description{
R isntallation folder needs to have standard annotation R-X.X.X
}
\examples{
\dontrun{
get_Rversion_from_path()
}
}
