% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_update.R
\name{get_user_input}
\alias{get_user_input}
\title{get user input necessary for update}
\usage{
get_user_input()
}
\value{
dir_ls list()
\describe{
  \item{path_new}{path of new R installation}
  \item{path_old}{path of old R installation}
  \item{path_miniCRAN}{path to miniCRAN repository}
  \item{libs_new}{packages in new R installation}
  \item{libs_old}{packages in old R installation}
  \item{R_vers_run}{R version currently running}
  \item{R_vers_new}{R version of new installation}
  \item{R_vers_old}{R version of old installation}
  \item{server}{ logical, is server environment}
  \item{can_internet}{logical, has internet connection}
  \item{miniCRAN}{logical, is miniCRAN supposed to be used}
}
}
\description{
used by update_from_old_inst() and update_new()
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
