% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkgs.R
\name{install_pkgs}
\alias{install_pkgs}
\title{Re-install Packages from Previous Installation}
\usage{
install_pkgs(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{character vector describing the location of R library
trees to search through}
}
\description{
Provided the location of the previous R installation's package
library, the function will attempt to re-install each of the
packages found. Packages are installed with \code{renv::install()}
and will be added to the 'renv' package cache.

Packages can be installed from GitHub, GitLab, BioConductor, and
any repository listed in \code{getOption("repos")}. This would typically
be from CRAN and any other secondary repositories that may be set.
}
\examples{
if (interactive()) {
  install_pkgs()
}
}
