\name{kate}
\alias{kate}
\title{Edit source file in kate (KDE advanced text editor)...}
\usage{kate(func, src.root=getOption("src.root"))
}
\description{Edit source file in kate (KDE advanced text editor)}
\value{nothing. The side effect is that an R source file is opened in an editor.}
\author{Mark Cowley, 26/2/07}
\arguments{\item{func}{a function name, or the function's code}
\item{src.root}{The parent folder of source code}
}

