% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_miniCRAN.R
\name{miniCRAN_archive}
\alias{miniCRAN_archive}
\title{archive current miniCRAN repository, execute on server}
\usage{
miniCRAN_archive(overwrite = F, FORCE = F)
}
\arguments{
\item{overwrite}{Default: F}

\item{FORCE}{force execution on desktop, Default: F}
}
\seealso{
\code{\link[stringr]{str_detect}},\code{\link[stringr]{str_split}}
}
