% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_update.R
\name{create_miniCRAN}
\alias{create_miniCRAN}
\title{create fresh miniCRAN from scratch}
\usage{
create_miniCRAN(overwrite = F, path = "c:/miniCRAN",
  CRAN_repos = "https://cran.rstudio.com")
}
\arguments{
\item{overwrite}{Default: F}

\item{path}{path to miniCRAN repository}
}
\description{
creates a fresh miniCRAN repository, usually not needed because
  a miniCRAN is already in place
}
