% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_miniCRAN.R
\name{miniCRAN_update_all_packages_from_miniCRAN}
\alias{miniCRAN_update_all_packages_from_miniCRAN}
\title{update all packages from miniCRAN, execute on server}
\usage{
miniCRAN_update_all_packages_from_miniCRAN(FORCE = F)
}
\arguments{
\item{FORCE}{force exection on desktop, Default: F}
}
