% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version.R
\name{r_version}
\alias{r_version}
\alias{previous_r_version}
\title{R Versions}
\usage{
r_version()

previous_r_version()
}
\value{
string of R version
}
\description{
These functions return the current R version and the
the previously installed R version.
}
\examples{
r_version()

previous_r_version()
}
