\name{smultron}
\alias{smultron}
\title{Edit source file in Smultron}
\usage{
  smultron(func, src.root = getOption("src.root"))
}
\arguments{
  \item{func}{a function name, or the function's code}

  \item{src.root}{The parent folder of source code}
}
\value{
  nothing. The side effect is that an R source file is
  opened in an editor.
}
\description{
  Edit source file in Smultron
}
\author{
  Mark Cowley, 26/2/07
}

