\name{trim}
\alias{trim}
\title{Trim...}
\usage{trim(x)
trim(x)
}
\description{Trim}
\details{\code{trim}: Trim the whitespace from front and back of the words in a vector.

\code{trim}: Trim the whitespace from front and back of the words in a vector.


}
\value{\code{trim}: character vector of trimmed text

\code{trim}: character vector of trimmed text


}
\author{Mark Cowley, 2009-08-19}
\arguments{\item{x}{[character] a character vector}
}
\alias{trim}

