% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_update.R
\name{update_from_old_inst}
\alias{update_from_old_inst}
\title{update new R installation based on old installation}
\usage{
update_from_old_inst(dir_ls = get_user_input())
}
\arguments{
\item{dir_ls}{list, Default: get_user_input()}
}
\description{
needs to run on old R version, copies all packages from old
  installation, updates Rprofile.site and archives miniCRAN if run on server
  environment.
}
