% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_update.R
\name{update_new_inst}
\alias{update_new_inst}
\title{update new R installation}
\usage{
update_new_inst(dir_ls = get_user_input())
}
\arguments{
\item{dir_ls}{PARAM_DESCRIPTION, Default: get_user_input()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
update new R installation, update all packages from either CRAN
  or miniCRAN. On desktop environment all packages not yet in miniCRAN are
  added. On server environment all missing packages from miniCRAN which are
  not yet installed will be installed.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[miniCRAN]{pkgAvail}},\code{\link[miniCRAN]{addPackage}}
}
