% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_check.R
\name{check_internet}
\alias{check_internet}
\title{check for an internet connection}
\usage{
check_internet(url = "http://www.google.com")
}
\arguments{
\item{url}{character vector, Default: 'http://www.google.com'}
}
\value{
logical
}
\description{
checks for an internet connection by reading the first line of a
  given url
}
\examples{
check_internet()
}
