% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_check.R
\name{check_libPaths}
\alias{check_libPaths}
\title{FUNCTION_TITLE}
\usage{
check_libPaths()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
