% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_check.R
\name{check_no_of_tasks}
\alias{check_no_of_tasks}
\title{check number of running tasks}
\usage{
check_no_of_tasks(task_regex = c("rsession\\\\.exe", "Rgui\\\\.exe",
  "R\\\\.exe", "Rscript\\\\.exe", "Rcmd\\\\.exe"))
}
\arguments{
\item{task_regex}{regex pattern for task, Default: 'rsession\.exe'}
}
\value{
integer, number of tasks
}
\description{
searches a sstring returned by the system command tasklist for a
  task regex pattern. Returns the number of occurrences.
}
\examples{
check_no_of_tasks()
}
