% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_miniCRAN.R
\name{miniCRAN_create}
\alias{miniCRAN_create}
\title{create fresh miniCRAN from scratch, not recommended}
\usage{
miniCRAN_create(overwrite = F, FORCE = F)
}
\arguments{
\item{overwrite}{Default: F}

\item{FORCE}{force execution if not in server environment, Default: F}
}
