% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater-package.R
\docType{package}
\name{updater-package}
\alias{updater}
\alias{updater-package}
\title{updater: Utilities When Updating R}
\description{
When updating major or minor R versions all packages should be re-installed. The utilities in this package assist in getting a user up-and-running again by installing all previously installed R packages. The package uses 'renv' to install; immediately replenishing your 'renv' package cache.
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [copyright holder]

}
\keyword{internal}
