\name{updateR}
\alias{updateR}
\title{updateR}
\usage{
updateR()
}
\description{
this function allows you to install the last version of R language on your OSX.
updateR takes care of detecting, downloading and installing last R version directly from CRAN.
no arguments or specifications ar needed, just run update()
}
\examples{
updater()
}
