% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{updateR}
\alias{updateR}
\title{Downloads and installs the latest version of R for Mac OS X.}
\usage{
updateR()
}
\description{
Update your version of R from inside R itself (Mac OS X only).
}
\examples{
\dontrun{
## NOT RUN
updateR()
## END NOT RUN
}
}
\author{
Andrea Cirillo, Robert Myles McDonnell
}
